+++ addons/ht_crm/models/customer/__init__.py
@@ -1,3 +1,4 @@
 from . import customer
 from . import phonebook
+from . import phonebook_batch
 from . import customer_report
